requirements.txt
dash
dash-bootstrap-components
PyMuPDF
textblob
wordcloud
pillow
pandas
plotly
